// src/schemas/API/trainingProgramResponse.ts
export interface trainingProgramResponse {
    trainingProgramId: string; // Sửa từ id thành trainingProgramId
    trainingProgramName: string;
    educationLevel: string;
    degreeType: string;
    trainingMode: string;
    trainingDuration: number;
    totalCredits: number;
    managingFaculty: string;
    teachingLanguage: string;
    status: boolean;
    website: string;
    issuingDecision: string;
    createdAt: string;
    updatedAt: string;
}